s/apparmor: include files in apparmor kernel features (#15130)
Previously, only directories under /sys/kernel/security/apparmor/features
were included in the probed kernel features list. However, boolean files
are sometimes used to indicate support for particular features as well.

This commit extends the probed features to include the presence of
files, as well as directories.

Though we are looking in a sysfs, /sys/kernel/security/apparmor/features
is entirely managed by apparmorfs, so it should not be a problem to
include files there in the kernel features (and by extension in the
system key).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>