s/a/notify: simplify the listener
No more `tomb`, no more waiting goroutines, no more atomics.

The listener run loop is now synchronous, and a `once.Do` ensures that
it only runs once, and thus the `reqs` channel is only closed once.

The `Reply` method on `Request` now directly encodes and sends the
response to the kernel. As a result, the `Request` struct now contains
the ID of the message from which it was derived, a pointer to the
listener, and the permission mask which was already allowed by existing
AppArmor rules.

A close channel is used to signal to the main loop that the listener has
been closed.

A `Mutex` guards the notify file, the epoll instance, and the close
channel, ensuring that the listener run loop and any replies cannot
race, and that there are no races around checking whether the listener
has been closed.

There is a new `MsgNotificationGeneric` interface which the mediation
class-specific message structs (e.g. `MsgNotificationFile`) should
implement, and which abstracts the behavior required of all mediation
classes and means the listener no longer has mediation class-specific
functions.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>