o/confdbstate: add helper for snapctl confdb load
Adds a GetTransactionForSnapctlGet helper that returns a Transaction for
`snapctl get` to read a confdb. If there are hooks to run, it schedules
the right tasks and tracks the transaction. This is meant only for
snapctl, the API will use a different helper since the circumstances
there differ quite a bit.
Also changes the tracking of the transaction to distinguish between read
and write transactions, since there can be many concurrent read
transaction but not write. A follow-up PR will use that to check for
conflicts earlier in the read/write.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>