i/prompting: use omitzero when encoding time.Time (#15112)
* i/prompting: use omitzero when encoding expiration

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: use omitzero instead of omitempty when encoding time.Time

Go 1.24 will support a new `omitzero` annotation for json encoding
struct fields. The existing `omitempty` does not consider time zero to
be empty, so where `omitempty` is used with non-pointer `time.Time`
values, the result is usually a marshalled time zero, rather than an
omitted field.

Using `omitzero` will fix this issue when on go 1.24+, and on earlier go
versions, `omitempty` did not function properly anyway, and `omitzero`
will be ignored.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: fix time.Time zero value expected omission in tests

If running on a Go version prior to 1.24, expect zero values of
`time.Time` to be encoded as `"0001-01-01T00:00:00Z"`, since
`"omitzero"` is ignored before version 1.24.

We previously used `"omitempty"` when encoding `time.Time`, but time
zero is not actually empty, so we ended up `"0001-01-01T00:00:00Z"`
anyway.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>