{i,s}/apparmor: add function to wrap apparmor snippet in metadata tags
AppArmor rules, be that individual rules or snippets, may be tagged with
metadata tags, as long as the AppArmor parser and kernel both support
tags. Tag blocks may also be nested.

This PR adds a function to check that tags are supported and a function
to wrap an AppArmor snippet in given tags.

When an AppArmor rule applies to a particular operation, the kernel can
provide the list of tags associated with that rule in any messages it
sends over the notify socket.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>