many: make API confdb reads load ephemeral data
Load ephemeral confdb data when reading from the API (for snap get).
This requires making the read side of the API async, since it might be
necessary to run hooks. The result in then transmitted through the
change to the client. In case there are no custodian hooks to run, no
tasks are run and the result is just directly placed into a "done"
change.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>