cmd: enforce application tracking for snaps using core26+
When snap applications or hooks are launched, they are tracked using a
systemd facility that requires cooperation from systemd running either
as root, or from systemd running as user (talking to systemd running as
root).

Since this feature was introduced, snapd had allowed for the tracking
functionality to fail silently. Now, when a snap application is using
core, core18, core20, core22 or core24 this behavior is retained, but
for snap applications based on core26 or newer, the failure is now
blocking application startup.

A message is printed to the user to inform them of this situation. The
message uses an abbreviated URL to the forum so that the message can be
edited and stay relevant over time.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-34299
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>