i/apparmor: adjust metadata tag formatting
Regardless of whether the snippet contains one or more rules, always
format tagged snippets as follows:

```
<non-tagged rule(s)>

tags=(foo,bar) {
<tagged rule(s)>
}

<other non-tagged rule(s)>
```

That is, always wrap the snippet to tag in newline characters inside the
tag `{` `}`, and additionally place a newline before and after the
tagged block as a whole, to better differentiate it from surrounding
rules. This is especially useful when nesting tagged snippets, so it's
easier to differentiate which tag block is closed by a given `}`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>