s/a/notify: register listener ID before setting filter
As part of `RegisterFileDescriptor`, register the listener ID with the
kernel before setting the filter. If there was a previous listener ID
saved on disk, send it in the message to tell the kernel to re-register
the existing listener to the file descriptor, otherwise the kernel will
populate the listener ID field with the ID of the new listener.
Regardless, save the listener ID to disk (in a tmpfs) so that snapd can
reclaim the existing listener after any future snapd restart.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>