core-initrd: load modules required by systemd
At this point UC24, does not have modules available in the start of
the main boot. This is a bug because modules could be loaded at any
time by the kernel.

Until this is fixed, we pre-load the kernel modules expected by
systemd to be loaded at start.

See https://github.com/systemd/systemd/blob/main/src/core/kmod-setup.c