s/a/notify: simplify the listener (#15178)
* s/a/notify: simplify the listener

No more `tomb`, no more waiting goroutines, no more atomics.

The listener run loop is now synchronous, and a `once.Do` ensures that
it only runs once, and thus the `reqs` channel is only closed once.

The `Reply` method on `Request` now directly encodes and sends the
response to the kernel. As a result, the `Request` struct now contains
the ID of the message from which it was derived, a pointer to the
listener, and the permission mask which was already allowed by existing
AppArmor rules.

A close channel is used to signal to the main loop that the listener has
been closed.

A `Mutex` guards the notify file, the epoll instance, and the close
channel, ensuring that the listener run loop and any replies cannot
race, and that there are no races around checking whether the listener
has been closed.

There is a new `MsgNotificationGeneric` interface which the mediation
class-specific message structs (e.g. `MsgNotificationFile`) should
implement, and which abstracts the behavior required of all mediation
classes and means the listener no longer has mediation class-specific
functions.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: adjust MsgNotificationGeneric methods and internal field names

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: adjust placement and documentation for MsgNotificationGeneric

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: improve documentation of close handling

Additionally, rename `runOnce` to `handleRequests`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: unexport the listener pointer embedded in requests

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: clarify purpose of once.Do and use more robust error check

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/i/apparmorprompting: clarify documentation around interaction with listener

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: clarify reason for buffered channel in mocked epoll/ioctl

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: make BuildResponse return a pointer

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: use int32 instead of uint32 for PID

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: rename Id field to KernelNotificationID

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: close the notify socket FD before closing the epoll

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: close notify socket FD before closing close channel

Close the notify socket FD to prevent the kernel from sending a new
request after we have already closed the close channel, which would
prevents that request from being handled properly.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: adjust wording of comments

Co-authored-by: Zygmunt Bazyli Krynicki <me@zygoon.pl>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Zygmunt Bazyli Krynicki <me@zygoon.pl>