s/a/n/listener: add ready method when all re-sent messages re-received
The kernel keeps track of two queues for each listener: the "ready"
queue and the "pending" queue. The "ready" queue is new notifications
which have yet to be read for the first time, and the "pending" queue is
notifications which have previously been read (presumably by the snapd
listener).

The snapd listener now keeps track of how many previously-sent messages
were pending when the listener was re-registered, and then for each
message received from the kernel which has the `NOTIF_RESENT` flag set,
that count is decreased. Once the listener has re-received the expected
number of messages with the flag set, then it will report itself as
ready.

The listener has a new `Ready()` method which returns a channel which is
closed when the listener becomes ready (which may be immediately, if
there are no pending notifications).

If the listener receives other messages before it has re-received all
messages it expects to be re-sent, it will queue those messages up and
wait to send them until it has become ready.

As a result, the requestprompts backend will (after some work) be able
to re-create the prompts state from prior to a snapd restart, and will
then be able to handle new requests, replies, and rules which have
occurred since then.

Assuming the kernel re-sends all pending notifications quickly, the
listener should become ready very soon after it is re-registered.
However, if it does not, or if any pending notifications time out in the
kernel before they are re-read by the listener, the listener will never
become ready. It may be necessary to add a timer-based override so that
the listener eventually reports itself as ready even if not all
notifications have been re-received.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>