o/i/apparmorprompting: ensure request handled before signalling ready
It's not sufficient to block method calls from the API until the
listener says the final pending request has been re-sent. Instead, the
method calls must block until the manager has handled, in addition to
received, the final request.

As such, the manager (not the listener) must be in control of the
channel used to unblock method calls from the API. The manager can wait
for the listener to ready in the same select where it waits for
requests. Requests synchronize before ready in the listener, so the
manager can ensure that it has handled the new request before it checks
again whether the listener is ready.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>