s/a/n/listener: signal ready after a timeout even if still waiting
Even if the listener is still waiting to re-receive previously-sent
requests which were originally received before snapd restarted, signal
that the listener is ready after a timeout.

This ensures that if a request times out in the kernel or is otherwise
dropped before it's re-received by the listener, that will not block the
listener from ever being ready.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>