cmd/snap-bootstrap: mount /lib/{modules,firmware} from snap as fallback
In case we do not have a drivers tree on 24+, mount /lib/{modules,firmware}
from the kernel snap as a fallback. This should not really happen and we do not
support fully this scenario, but there are cases where this mihgt be necessary,
like if somebody uses a core22 model but install a 24+ kernel.