s/a/n/listener: ensure thread safety around internal readiness checks
Use a mutex to protect `pendingCount` and `readyQueue`, as both of these
are written by both the main run loop and the `signalReadyAndFlushQueue`
method, which may be triggered by a timer expiration callback goroutine.

Additionally, fix a bug where `pendingCount` was not zeroed out by
`signalReadyAndFlushQueue`, so if the ready timer timed out, the main
run loop would still attempt to queue non-`RESENT` requests.

Lastly, add a dedicated `doneReadying` channel which is closed after
`signalReadyAndFlushQueue` is done writing to `l.reqs`, so that the
cleanup when the main run loop ends can safely wait to close `l.reqs`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>