{i,s}/apparmor: add function to wrap apparmor snippet in metadata tags (#15199)
* {i,s}/apparmor: add function to wrap apparmor snippet in metadata tags

AppArmor rules, be that individual rules or snippets, may be tagged with
metadata tags, as long as the AppArmor parser and kernel both support
tags. Tag blocks may also be nested.

This PR adds a function to check that tags are supported and a function
to wrap an AppArmor snippet in given tags.

When an AppArmor rule applies to a particular operation, the kernel can
provide the list of tags associated with that rule in any messages it
sends over the notify socket.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/apparmor: add MetadataTag type with unique registration

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/apparmor: adjust metadata tag formatting

Regardless of whether the snippet contains one or more rules, always
format tagged snippets as follows:

```
<non-tagged rule(s)>

tags=(foo,bar) {
<tagged rule(s)>
}

<other non-tagged rule(s)>
```

That is, always wrap the snippet to tag in newline characters inside the
tag `{` `}`, and additionally place a newline before and after the
tagged block as a whole, to better differentiate it from surrounding
rules. This is especially useful when nesting tagged snippets, so it's
easier to differentiate which tag block is closed by a given `}`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/apparmor: remove unused error return value from MetadataTagSnippet

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/apparmor: remove tag registration without interface

There's not a well-defined use case for registering a tag as
unassociated with any interface, so remove the ability to do so.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/apparmor: update comment to remove reference to unassociated tag registration

Co-authored-by: Maciej Borzecki <maciek.borzecki@gmail.com>

* i/apparmor: restrict metadata tag format and use space as separator

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Maciej Borzecki <maciek.borzecki@gmail.com>