daemon,o/restart: add method to check pending restart without state lock
In places such as the daemon, we acquire state lock just to check
whether a restart is pending. This increases state lock contention and
slows down API replies.

Thus, make the restart value atomic, allowing it to be checked via a
method on the `RestartManager` without needing to lock snapd state.

This is an adaptation of a similar change from pebble:

https://github.com/canonical/pebble/pull/451

Because many other places do not have a reference to an overlord from
which they call `RestartManager()` to get access to the manager, we also
preserve the existing `restart.Pending()` function, which differs from
the change in pebble.

Snapd is still on go 1.18, so we cannot yet use `atomic.Int32` as pebble
does. Instead, we use `atomic.{Load,Store,Swap}` on an `int32`, but we
should switch to the atomic types once snapd's go version is advanced.

Lastly, there are a few other drive-by changes identified in snapd which
were not previously included in pebble. The following PR ports them back
to pebble: https://github.com/canonical/pebble/pull/597

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>