bash-completion: Add some bash completion to run command
The run command only offers bash completion for their own
parameters, but not for the snap application to run.

This patch adds autocompletion support for that. It allows to
autocomplete the snap name by default, listing only those that
have applications inside (thus avoiding to list base snaps like
coreXX). Also, by deleting the space and adding a dot, will
complete the application names in the specified snap.