bash-completion: Add some bash completion to run command
The run command only offers bash completion for the parameters,
but not for the snap application to run.

This patch adds some (incomplete) bash completion support for
that, by using the folder list in /snap as source for completion
when asking for a program. It is not perfect, because it's not
possible to autocomplete other apps inside a snap, but is better
than the current situation.