s/a/notify: add MetadataTags and TagsetMap types
Add a `MetadataTags` type which is a list of metadata tags in the order
they were received from the kernel. This can replace usage of `[]string`
in various places in the package.

Add a `TagsetMap` type which maps from AppArmor permission to
`MetadataTags`. This matches the format of metadata tags as received
from AppArmor in the kernel, and is used in place of the previous map
stored in `MsgNotificationFile` and `packTagsets`.

The `TagsetMap` type is defined with the `InterfaceForPermission`
method, which computes the interface based on the metadata tags which
apply to the given permission mask. If more than one tagset overlaps
with the given permission mask, then the tags in those tagsets are
concatenated. The interface is computed by attempting to look up the
registered interface for each tag in the tags list, and for the first
which has a registered interface, that interface is returned.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>