s/a/n/listener: add Interface field to Request and populate from tags
Add an `Interface` field to `listener.Request`, and populate it via the
`Interface()` method on the received `MsgNotificationGeneric`. If no
tags are present, or none received are registered with an interface,
then `Interface()` returns false, and the listener defaults to using the
`"home"` interface, since prompting without metadata tags set is only
used with the home interface.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>