i/p/requestprompts: make change to tmpfs dir backwards compatible
The request prompt max ID file is moved from snap run dir to the
`interfaces-requests` directory within the snap run dir.

To make this change backwards compatible, we first check if the max ID
file does not exist, but the legacy file does, and if so, hardlink the
legacy file to the new location.

That way, snapd can always open the mmap at the new location, and if
snapd is reverted (prior to a system restart), the max ID file is still
present at the legacy location and has continued to be updated as the
max ID was incremented.

We don't guarantee backwards compatibility if the user reverts snapd
after they've restarted their system, since doing so would require that
we always have the file in both locations (hardlinked). And again, this
is the tmpfs where prompt IDs are stored, and we don't expect prompts to
be long-lived, so it should be very difficult to hit this in the real
world. The result would be re-using prompt IDs, which if the earlier
prompt is still pending, would result in the prompt being lost. This
would be bad, but again, this should be exceedingly difficult to hit.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>