osutil: add vfs package
The vfs package provides a virtual file system sufficient for
snap-update-ns to run two simulations of a sequence of mount operations
and use Stat to determine if two paths are really pointing to the same
resource.

This is phase one, as phase two is also needed to provide recursive bind
mounts on top of the core functionality.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-34727

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>