tests: new workflow to check state locks against the baseline (#15314)
* tests: new workflow to check state locks againt the baseline

This workflow runs snapd tests and check there is not new state lock
traces comparing with the baseline.

The workflow builds snapd with testkeys to allow state locks are
generated, and the run the main and core test suites to obtain the locks
generated (using a threshold of 1000 ms).

Once the artifacts are generated and retrieved, the lock files are
compiled into a single log and the traces are compared against the
baseline.

If new traces are found in the logs, the workflow will print the diff
and fail.

State lock logs are always uploaded as an artifact.

* use snapd from master as it is build with tag statelocktrace