o/confdbstate: add early concurrency checks  (#15305)
* o/confdbstate: add early concurrency checks

Add early concurrency checks to read/write confdb operations. These
checks also happen when an operation is registered but the intent here
is to fail (eventually to be replaced with blocking) before any
change/transaction is created.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* o/confdbstate: move missing assertion test

Move the missing assertion test from ctlcmd to confdbstate which is
easier to test and more appropriate.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: add a view.ID helper

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* o/confdbstate: add helpers

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>