pipewire: Add access to impl.portal.PermissionStore
Wireplumber requires access to the PermissionStore DBus service,
so access to it must be added to the slot side of the pipewire
interface.

For the moment, the "peer=(label=unconfined)" is used, since
the PermissionStore daemon is still running unconfined; in the
future, since it will probably require its own interface, an
specific peer rule will be added.