many: rewrite interface selection and move to i/prompting
Instead of concatenating the metadata tags from all tagsets into one
list and then returning the first interface registered in association
with one of those tags, instead take a more cautious approach.

When we receive a message from the kernel, we need to create a request
from it which has a single interface. Thus, we need to choose an
interface which applies to every permission requested in the message.
That is, every permission which was denied in the original message.

First, filter the tagsets to only include those associated with
requested permissions -- this is the responsibility of s/a/notify.
Then, include that filtered tagset map in the `listener.Request`.

Then, when `o/i/apparmorprompting` receives the new request from the
listener, it needs to choose an interface to use based on the tags. That
decision is delegated to a helper function in `i/prompting`, since it
needs to call functions in `i/apparmor` to look up any interface
associated with a given tag.

That helper function returns `ErrNoInterfaceTags` if there are no
tagsets, or if none of the tags in any tagset is registered in
association with an interface. If this is the case, then the
apparmorprompting manager defaults to the "home" interface.

Otherwise, the helper either returns an interface which applies to every
requested permission, or it returns a different error. Tags which are
not registered in association with an interface are ignored. An
interface applies to a permission if a tag which occurs in a tagset for
a permission mask which contains that permission was registered in
association with that interface.

If more than one interface applies to every requested permission, then
the helper prioritizes the interface associated with the most "specific"
tag, on average. That is (after discarding unregistered tags), the tag
which occurs closest to the end of each tagset, on average. The kernel
orders tags in the tagset by their specificity to a particular rule, and
snapd preserves that order.

In the unlikely event that more than one interface occurs on average the
same distance from the end of the tagsets, then one is chosen
arbitrarily (but consistently). This should never occur in practice.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>