i/prompting: simplify interface selection based on metadata tags
In practice, we should only ever see tags which are associated with a
single interface for any given request, and if we see any tags are
associated with an interface, those tags should be present in every
tagset, and thus apply to every requested permission.

Thus, the interface selection process can be greatly simplified.

1. If more than one interface is found to be associated with tags in the
   request, then return an error.
2. If an interface is found to be associated with tags, but at least one
   permission has no tags associated with that interface, then return an
   error.

No more interface selection based on the "most specific" tags associated
with an interface.

If there comes a time when prompting is enabled for multiple interfaces
which grant access to overlapping filepaths, then this should be
reconsidered.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>