o/configstate/configcore: start/stop ssh.socket for 24+
Start/stop the ssh.socket unit instead of ssh.service when enabling/disabling
the ssh service (service.ssh.disable service option), where the unit
exists (that is, for UC24+). With this we prevent that on stop the socket unit
tries to retrigger the service (which makes it fail by retries), and on start
we do not start a sshd process that will be instead started by the socket unit
when needed.