s/a/n/listener: assume kernel sends resent messages first
The kernel can guarantee that after a RESEND command is received, the
pending queue is inserted at the start of the ready queue, thus
guaranteeing that any previously-sent request will be re-sent again
before any new requests are sent.

As a result, the listener can get rid of its ready queue and simplify
much of the readiness handling. If the listener sees a request without
the `UNOTIF_RESENT` flag set, then it knows the kernel is done
re-sending previously-sent requests, and it should ready up. There is no
need to queue requests anymore.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>