s/a/notify: select interface from metadata tags and set in request (#15311)
* s/a/notify: add MetadataTags and TagsetMap types

Add a `MetadataTags` type which is a list of metadata tags in the order
they were received from the kernel. This can replace usage of `[]string`
in various places in the package.

Add a `TagsetMap` type which maps from AppArmor permission to
`MetadataTags`. This matches the format of metadata tags as received
from AppArmor in the kernel, and is used in place of the previous map
stored in `MsgNotificationFile` and `packTagsets`.

The `TagsetMap` type is defined with the `InterfaceForPermission`
method, which computes the interface based on the metadata tags which
apply to the given permission mask. If more than one tagset overlaps
with the given permission mask, then the tags in those tagsets are
concatenated. The interface is computed by attempting to look up the
registered interface for each tag in the tags list, and for the first
which has a registered interface, that interface is returned.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: add Interface field to Request and populate from tags

Add an `Interface` field to `listener.Request`, and populate it via the
`Interface()` method on the received `MsgNotificationGeneric`. If no
tags are present, or none received are registered with an interface,
then `Interface()` returns false, and the listener defaults to using the
`"home"` interface, since prompting without metadata tags set is only
used with the home interface.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/i/apparmorprompting: get interface from incoming request

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: rewrite interface selection and move to i/prompting

Instead of concatenating the metadata tags from all tagsets into one
list and then returning the first interface registered in association
with one of those tags, instead take a more cautious approach.

When we receive a message from the kernel, we need to create a request
from it which has a single interface. Thus, we need to choose an
interface which applies to every permission requested in the message.
That is, every permission which was denied in the original message.

First, filter the tagsets to only include those associated with
requested permissions -- this is the responsibility of s/a/notify.
Then, include that filtered tagset map in the `listener.Request`.

Then, when `o/i/apparmorprompting` receives the new request from the
listener, it needs to choose an interface to use based on the tags. That
decision is delegated to a helper function in `i/prompting`, since it
needs to call functions in `i/apparmor` to look up any interface
associated with a given tag.

That helper function returns `ErrNoInterfaceTags` if there are no
tagsets, or if none of the tags in any tagset is registered in
association with an interface. If this is the case, then the
apparmorprompting manager defaults to the "home" interface.

Otherwise, the helper either returns an interface which applies to every
requested permission, or it returns a different error. Tags which are
not registered in association with an interface are ignored. An
interface applies to a permission if a tag which occurs in a tagset for
a permission mask which contains that permission was registered in
association with that interface.

If more than one interface applies to every requested permission, then
the helper prioritizes the interface associated with the most "specific"
tag, on average. That is (after discarding unregistered tags), the tag
which occurs closest to the end of each tagset, on average. The kernel
orders tags in the tagset by their specificity to a particular rule, and
snapd preserves that order.

In the unlikely event that more than one interface occurs on average the
same distance from the end of the tagsets, then one is chosen
arbitrarily (but consistently). This should never occur in practice.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting: simplify interface selection based on metadata tags

In practice, we should only ever see tags which are associated with a
single interface for any given request, and if we see any tags are
associated with an interface, those tags should be present in every
tagset, and thus apply to every requested permission.

Thus, the interface selection process can be greatly simplified.

1. If more than one interface is found to be associated with tags in the
   request, then return an error.
2. If an interface is found to be associated with tags, but at least one
   permission has no tags associated with that interface, then return an
   error.

No more interface selection based on the "most specific" tags associated
with an interface.

If there comes a time when prompting is enabled for multiple interfaces
which grant access to overlapping filepaths, then this should be
reconsidered.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: add test of MetadataTagsets method

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: address review comments around interface selection

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: improve metadata tagsets documentation and consistency across mediation classes

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/notify: rename MetadataTagsets to DeniedMetadataTagsets

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>