s/a/notify: add structs/commands for v5 listener registration (#15387)
* s/a/notify: add structs/commands for v5 listener registration

In order to support snapd reclaiming outstanding notifications after it
restarts, we need to register the listener ID with the kernel and
re-register it after a restart. Then, we send a request to re-send
outstanding messages.

This commit adds the ioctl register and resend commands and defines the
structs used for those commands.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/a/n/listener: clarify default use of flags in tests

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>