packaging/arch: simplify PKGBUILD, sync with downstream
The PKGBUILD file was originally based on one for snapd-git AUR package.
However, since then snapd-git became defunct, at the same time all of
the downstream packaging is done in the snapd AUR package.

The patch syncs the packaging so that it's almost identical, with a
small exception of the header and lack of patches.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>