packaging/arch, tests/lib/prepare-restore: simplify PKGBUILD, simplify building of Arch package in CI (#15393)
* packaging/arch: simplify PKGBUILD, sync with downstream

The PKGBUILD file was originally based on one for snapd-git AUR package.
However, since then snapd-git became defunct, at the same time all of
the downstream packaging is done in the snapd AUR package.

The patch syncs the packaging so that it's almost identical, with a
small exception of the header and lack of patches.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/lib/prepare-restore: simplify packing of source tarballs for Arch

The PKGBUILD for Arch has been simplified and is more in line with how
the actual downstream packaging looks like. This allows us to simplify
packing of source tarballs, i.e. reuse pack-source script and drop the
custom sed/awk processing.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>