i/prompting: implement Text(Un)Marshaler for IDType
Instead of implementing `json.{M,Unm}arshaler`, implement the
`encoding.Text{M,Unm}arshaler` interface. `json.Marshal` uses the
text (un)marshaler interface as a fallback when the json (un)marshaler
interfaces are not implemented.

Importantly, when marshaling `IDType` to json as a key in a map, the
implementation of the json interfaces is not sufficient for the ID to be
properly marshaled in its string format; instead, the default `uint64`
marshaling is used. This is because json keys must be strings, and the
keys must be unique, so keys are marshalled to text first, then later
marshalled as json. As a result, by implementing the `TextMarshaler`
interface, the text representation of `IDType` is correct, and the json
marshaling inherits it.

Additionally, the `MarshalText` method must be implemented on a
non-pointer `IDType` receiver in order for `IDType` to be marshalled
properly as keys or values in maps. This is because keys and values in
maps are not addressible (see: `reflect.CanAddr`). Implementing
`MarshalText` on the literal `IDType` allows them to be marshalled as
keys/values.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>