i/p/requestprompts: add function to prune request map when ready occurs
When the listener is ready, any requests in the ID map which have not
yet been re-received from the kernel can be discarded from the map,
since readying indicates that there are no more pending requests waiting
to be re-received.

In doing so, the requestprompts backend should also record a notice for
any prompt IDs in the map for which no mapped requests have been
re-received, and thus the corresponding prompt has not been re-created.

If snapd restarts before the ready is handled, then we want to preserve
all IDs in the map, so we can handle all requests which may be received
after the restart.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>