o/i/apparmorprompting: ready prompt backend after listener readies
When the listener signals that all pending requests have been
re-received, and the manager has finished handling all those requests,
the manager should tell the prompts backend that it is ready. This
allows it to prune any unreceived requests from the ID map.

The manager should wait for the prompts backend to finish handling the
readying before it closes the manager-level ready channel, so that the
prompts backend is consistent before any potential API requests can hit
it.

The manager should only tell the prompts backend to ready if the
listener has signalled readiness, *not* if the manager is closing before
this occurs. This is because snapd is restarting quickly before all
requests have been re-received, and we want the opportunity to retry
re-receiving and handling all pending requests after snapd finishes
restarting.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>