i/p/requestprompts: deduplicate listener requests associated with prompt
Ensure that if a request is received multiple times (such as because
`APPARMOR_NOTIF_RESEND` is called again), a response with that request
ID is not sent multiple times, and the request is not associated
multiple times with the prompt.

This is important to avoid the kernel returning an error to repeated
replies against the same request ID, and so that we don't potentially
try to access and delete the same request ID from the ID map multiple
times.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>