many: non-setuid snap-confine, caps v4 (#15094)
* cmd/snap-confine: update cgroup code to run under an ordinary user

Since snap-confine is eventually going to be run under an ordinary user,
make sure that the snaps' cgroup directories are created as the root
user and group. Also, since we do not have direct control over the
ownership of the leaf files defining the cgroup (they are automatically
created using the effective user of the process that creates the
cgroup), remove the execute permission from the parent directory, so
that only the root user will be able to alter the snap's cgroup
configuration.

* snap-confine: create helper function to set capabilities

Create a wrapper that works well with 64 bit capabilities.

* libsnap-confine-private: add more capability-related methods

* cmd/snap-confine: drop setuid and use capabilities instead

* snap-confine: get the current working directory as a user

Read the process state (for the time being, that's just the current
working directory) as the real user who started the application, because
if we do this operation while we are setuid root, it would fail on some
filesystems (like SSHFS and NFS) which restrict access to the root user.

* cmd/snap-confine: add a few debug checkpoints for capabilities

* snap-confine: update apparmor profile for capability support

* cmd/snap-update-ns: explicitly specify ownership of mount profile

We'll eventually run snap-update-ns as a normal user (when invoked from
snap-confine, at least), so make sure that the mount profile continues
to be owned by root.root

* snap-confine: move AppArmor check earlier

Move the AppArmor check up, right after checking that the effective user
is root. Remove the check on the effective user being 0, since we assert
it right above.
Avoid calling getuid() and use the freshly obtained `real_uid` instead.

* snap-confine: remove calls to sc_set_effective_identity()

When needed, add code to explicitly adjust the group ownership.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* packaging: set capabilities

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/make: setcap capabilities when using 'hack' target

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/make: install-exec-hook is no longer needed

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine: raise root privileges before creating the cgroup freezer, otherwise this will fail under lxd

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* snapcraft: set snap-confine capabilities

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

snapcraft: require libcap tools

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/snap-confine-caps: verify snap-confine file capabilities

Add test which verifies snap-confine file capabilities.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* github: snap-builds: disable xattrs when extracing from snapd snap

Disable xattrs so that we do no error due to file capabilities.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/configure: check whether libcap supports cap_set_ambient

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd: provide alternative cap_set_ambient, cap_reset_ambient

When buildin on older systems like Ubuntu 18.04, the libcap library
version available there does not have
cap_set_ambient()/cap_reset_ambient(). Provide alternative
implementations based on the upstream sources.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private/tool: no identity change, use tool specific caps

There's no longer a need to switch identity when forking into
snap-update-ns. Privileges are transferred as ambient capabilities.

Make sure to carry only a limited/required set of capabilities when
invoking specific tools.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine: consolidate on using libcap and smart capability handling at runtime

Introduce smarter capability handling at runtime. We expect the
snap-confine to have file capabilities defining only the permitted set.
The effective capabilities are changed at runtime as needed.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* data/selinux: update SELinux policy

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-update-ns/bootstrap: verify capabilities, no uid switching

Do not switch uid/gid during bootstrap, as s-u-n will be invoked as a
target user. Verify that when running, CAP_SYS_ADMIN is effective.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private/privs: simplify dropping privileges

Use CAP_MODE_NOPRIV which is more strict than what was set earlier.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/snap-confine-undesired-mode-group: fix wording and checks

The test is not looking for writable files, but rather directories which
have g+x or o+x.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private/privs: drop unused helpers

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine: sc_ensure_mkdir() TODO

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private: use sc_ensure_mkdir() helpers, fix ownership

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine: join cgroup v1 freezer as user

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/lib/pkgdb: install libcap-progs on opensuse

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private: drop uses of sc_identity in tool

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine/mount-support: comment cases which do not need mode 0000

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine: drop CAP_SETUID, CAP_SETGID, no longer needed

Since we no longer switch the euid/egid, CAP_SETUID & CAP_SETGID are not
needed.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* libsnap-confine-private/privs: drop sc_set_keep_caps_flag()

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap-confine: do not repeat a static string

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private/locking: comment on not needing 000

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* many: use a central file with file capabilities

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>
Co-authored-by: Alberto Mardegan <alberto.mardegan@canonical.com>