asserts/signtool: format JSON in assertion bodies
In order to keep assertions readable and consistent, we require the JSON
in an assertion body (e.g., confdb-schema) to be indented with two
spaces per level and have all object keys sorted.
This change makes the signtool reformat assertion bodies. This makes the
UX of writing and signing an assertion (whether with snapcraft or snap
sign) much better, since it can be difficult to get the format right "by
hand".

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>