s/a/notify: add handling for kernel listener timeout and more logging
The kernel will return `ENOENT` on `APPARMOR_NOTIF_REGISTER` if the
listener with the given ID does not exist, such as if the listener
timed out. This PR adds handling for snapd to remove the saved listener
ID and retry if it tries to register the listener with a saved ID and
gets a `ENOENT` instead.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>